refactor(telegram): extract s2 polling/dm-topics clusters into mixin modules (epic #78791)
